requirements.txt
streamlit>=1.32.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.2.0
langdetect>=1.0.9
pypdf>=4.0.0
python-docx>=1.1.0

pytest>=8.0.0
pytest-asyncio>=0.23.0
ruff>=0.5.0
black>=24.0.0
mypy>=1.8.0
